requirements.txt
asgiref==3.8.1
boto3==1.34.122
botocore==1.34.122
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
django-storages==1.14.3
djangorestframework==3.15.1
jmespath==1.0.1
Markdown==3.6
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
